Add complete payment endpoint

+++ teacup-loan-payments/src/main/java/com/example/teacuploanpayments/entity/PaymentRepository.java
@@ -2,5 +2,5 @@ package com.example.teacuploanpayments.entity;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PaymentRepository extends CrudRepository<Payment, Long> {
+public interface PaymentRepository extends CrudRepository<Payment, Integer> {
 }
